Update error handling method.
`webpack.NoEmitOnErrorsPlugin()` now appears to cause the tooling to
lock up. Switching to a more direct handling of webpack failures here.
Tapping into the `done` hook and checking for errors. If any exist,
print out information, and hard exit the process with an error code.
Throwing an error here instead would cause `handleBuildResult` in
`karma-webpack` to fail to resolve a callback and lock up, or fail in
other ways with naive fixes to that code path.